Create README.md
I'm doing this project under guidence of Dr. Tushar Sandhan (Dept. of Electrical Enginneering), all credits for this project goes to him. We have data of more than 5 lakh invoice of a E-commerce site we segmented the customers aiming to assist organizations in understanding their customers and making strategic decisions for product grow.